Moves application code to src folder and updates imports
