+++ package.json
@@ -2,9 +2,9 @@
   "name": "quizzer-fullstack",
   "version": "1.0.0",
   "description": "Web application for quizzes",
-  "main": "index.js",
+  "main": "./src/index.js",
   "scripts": {
-    "dev": "NODE_ENV=dev node index.js",
+    "dev": "NODE_ENV=dev node ./src/index.js",
     "test": "NODE_ENV=test jest --verbose --runInBand",
     "format:check": "prettier --check .",
     "format:write": "prettier --write .",
