+++ tests/app.test.js
@@ -1,5 +1,5 @@
 const supertest = require('supertest')
-const app = require('../app')
+const app = require('../src/app')
 
 const API = supertest(app)
